# swagger.json
{"openapi":"3.0.0","paths":{"/session/login":{"post":{"operationId":"SessionController_postLogin","summary":"Post the login credentials data","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/PostLoginReqDto"}}}},"responses":{"200":{"description":"Returns a JSON with the jwt data","content":{"application/json":{"schema":{"$ref":"#/components/schemas/PostLoginResDto"}}}},"500":{"description":"Internal server error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ErrorDto"}}}}},"tags":["session"]}},"/user/get":{"get":{"operationId":"UserController_getUser","summary":"Get the user data","parameters":[],"responses":{"200":{"description":"Returns a JSON with the user data","content":{"application/json":{"schema":{"$ref":"#/components/schemas/GetUserResDto"}}}},"500":{"description":"Internal server error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ErrorDto"}}}}},"tags":["user"],"security":[{"auth":[]}]}},"/user/post":{"post":{"operationId":"UserController_postUser","summary":"Post the user data","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/PostUserReqDto"}}}},"responses":{"200":{"description":"Returns a JSON with the user data","content":{"application/json":{"schema":{"$ref":"#/components/schemas/GetUserResDto"}}}},"500":{"description":"Internal server error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ErrorDto"}}}}},"tags":["user"]}},"/user/put":{"put":{"operationId":"UserController_putUser","summary":"Put the user data","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/PutUserReqDto"}}}},"responses":{"200":{"description":"Returns a JSON with the user data","content":{"application/json":{"schema":{"$ref":"#/components/schemas/GetUserResDto"}}}},"500":{"description":"Internal server error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ErrorDto"}}}}},"tags":["user"],"security":[{"auth":[]}]}},"/user/delete":{"delete":{"operationId":"UserController_deleteUser","summary":"Delete the user data","parameters":[],"responses":{"200":{"description":"Returns a JSON with the user status","content":{"application/json":{"schema":{"$ref":"#/components/schemas/DeleteUserResDto"}}}},"500":{"description":"Internal server error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ErrorDto"}}}}},"tags":["user"],"security":[{"auth":[]}]}},"/event/get/all":{"get":{"operationId":"EventController_getEvent","summary":"Get all events data","parameters":[],"responses":{"200":{"description":"Returns a JSON with all the events data","content":{"application/json":{"schema":{"$ref":"#/components/schemas/GetEventResDto"}}}},"500":{"description":"Internal server error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ErrorDto"}}}}},"tags":["event"]}},"/event/post":{"post":{"operationId":"EventController_postEvent","summary":"Post the event data","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/PostEventReqDto"}}}},"responses":{"200":{"description":"Returns a JSON with the event data","content":{"application/json":{"schema":{"$ref":"#/components/schemas/GetEventResDto"}}}},"500":{"description":"Internal server error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ErrorDto"}}}}},"tags":["event"],"security":[{"auth":[]}]}},"/event/put":{"put":{"operationId":"EventController_putEvent","summary":"Put the event data","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/PutEventReqDto"}}}},"responses":{"200":{"description":"Returns a JSON with the event data","content":{"application/json":{"schema":{"$ref":"#/components/schemas/GetEventResDto"}}}},"500":{"description":"Internal server error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ErrorDto"}}}}},"tags":["event"],"security":[{"auth":[]}]}},"/event/delete":{"delete":{"operationId":"EventController_deleteEvent","summary":"Delete the event data","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"number"}}],"responses":{"200":{"description":"Returns a JSON with the event status","content":{"application/json":{"schema":{"$ref":"#/components/schemas/DeleteEventResDto"}}}},"500":{"description":"Internal server error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ErrorDto"}}}}},"tags":["event"],"security":[{"auth":[]}]}}},"info":{"title":"tde-ms-event","description":"events","version":"1.0.0","contact":{}},"tags":[],"servers":[{"url":"/tde-ms-event"}],"components":{"securitySchemes":{"auth":{"scheme":"bearer","bearerFormat":"JWT","type":"http","name":"JWT","in":"header"}},"schemas":{"PostLoginReqDto":{"type":"object","properties":{"email":{"type":"string"},"password":{"type":"string"}},"required":["email","password"]},"PostLoginResDto":{"type":"object","properties":{"jwt":{"type":"string"}},"required":["jwt"]},"ErrorDto":{"type":"object","properties":{"statusCode":{"type":"number"},"message":{"type":"string"}},"required":["statusCode","message"]},"UserDto":{"type":"object","properties":{"name":{"type":"string"},"email":{"type":"string"},"isAdmin":{"type":"boolean"}},"required":["name","email","isAdmin"]},"GetUserResDto":{"type":"object","properties":{"user":{"$ref":"#/components/schemas/UserDto"}},"required":["user"]},"PostUserReqDto":{"type":"object","properties":{"name":{"type":"string"},"email":{"type":"string"},"password":{"type":"string"},"isAdmin":{"type":"boolean"}},"required":["name","email","password","isAdmin"]},"PutUserReqDto":{"type":"object","properties":{"name":{"type":"string"},"password":{"type":"string"}},"required":["name","password"]},"DeleteUserResDto":{"type":"object","properties":{"statusCode":{"type":"number"},"message":{"type":"string"}},"required":["statusCode","message"]},"GetEventResDto":{"type":"object","properties":{"event":{"type":"array","items":{"type":"string"}}},"required":["event"]},"PostEventReqDto":{"type":"object","properties":{"name":{"type":"string"},"eventDate":{"type":"string"},"location":{"type":"string"},"description":{"type":"string"}},"required":["name","eventDate","location","description"]},"PutEventReqDto":{"type":"object","properties":{"id":{"type":"number"},"name":{"type":"string"},"eventDate":{"type":"string"},"location":{"type":"string"},"description":{"type":"string"}},"required":["id","name","eventDate","location","description"]},"DeleteEventResDto":{"type":"object","properties":{"statusCode":{"type":"number"},"message":{"type":"string"}},"required":["statusCode","message"]}}}}